Move decode_sum_type to cbor::hs::util
